API reference

Index

ImplicitDifferentiation.Conditions

ImplicitDifferentiation.Conditions

ImplicitDifferentiation.Forward

ImplicitDifferentiation.Forward

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.PullbackMul!

ImplicitDifferentiation.PushforwardMul!

Docstrings

ImplicitDifferentiation.Conditions — Type

Conditions{handle_byproduct,C}

Callable wrapper for conditions c::C, which ensures that a byproduct z is always accepted in addition to x and y.

The type parameter handle_byproduct is a boolean stating whether or not c natively accepts z.

ImplicitDifferentiation.Conditions — Method

conditions(x, y, z; kwargs...)

Apply conditions.c to (x, y, z), discarding z beforehand if needed.

ImplicitDifferentiation.Forward — Type

Forward{handle_byproduct,F}

Callable wrapper for a forward mapping f::F, which ensures that a byproduct z(x) is always returned in addition to y(x).

The type parameter handle_byproduct is a boolean stating whether or not f natively returns z(x).

ImplicitDifferentiation.Forward — Method

forward(x; kwargs...)

Apply forward.f to x, returning a dummy byproduct z(x)=0 if needed.

ImplicitDifferentiation.ImplicitFunction — Type

ImplicitFunction{handle_byproduct,FF<:Forward,CC<:Conditions,LS}

Differentiable wrapper for an implicit function defined by a forward mapping and a set of conditions.

Details

If handle_byproduct=false, the forward mapping is x -> y(x) and the conditions are c(x,y(x)) = 0.

If handle_byproduct=true, the forward mapping is x -> (y(x),z(x)) and the conditions are c(x,y(x),z(x)) = 0. In this case, z(x) contains additional information generated by the forward mapping, but considered constant for differentiation purposes.

Given x ∈ ℝⁿ and y ∈ ℝᵈ, we need as many conditions as output dimensions: c(x,y,z) ∈ ℝᵈ. We can then compute the Jacobian of y(⋅) using the implicit function theorem:

∂₂c(x,y(x),z(x)) * ∂y(x) = -∂₁c(x,y(x),z(x))

This requires solving a linear system A * J = -B, where A ∈ ℝᵈˣᵈ, B ∈ ℝᵈˣⁿ and J ∈ ℝᵈˣⁿ.

Fields

forward::FF: a wrapper of type Forward coherent with the value of handle_byproduct

conditions::FF: a wrapper of type Conditions coherent with the value of handle_byproduct

linear_solver::LS: a callable of the form (A,b) -> (u,stats) such that Au = b and stats.solved ∈ {true,false}, typically taken from Krylov.jl

Constructors

ImplicitFunction(f, c; linear_solver=gmres)
ImplicitFunction(f, c, Val(handle_byproduct); linear_solver=gmres)

Construct an ImplicitFunction by wrapping a forward mapping f into a field of type Forward and conditions c into a field of type Conditions, taking into account the value of handle_byproduct (which defaults to false). The default linear solver is Krylov.gmres, but this can be changed with a keyword argument.

ImplicitDifferentiation.ImplicitFunction — Method

implicit(x; kwargs...)
implicit(x, Val(return_byproduct), ; kwargs...)

Make an ImplicitFunction callable by applying the forward mapping implicit.forward.

If return_byproduct=false (the default), this returns a single output y(x).

If return_byproduct=true, this returns a couple of outputs (y(x),z(x)).

The argument return_byproduct is independent from the type parameter handle_byproduct in ImplicitFunction, so any combination is possible.

ImplicitDifferentiation.PullbackMul! — Type

PullbackMul!{P,N}

Callable structure wrapping a pullback with N-dimensional outputs into an in-place multiplication for vectors.

Fields

pullback::P: the pullback of the function

output_size::NTuple{N,Int}: the array size of the function output

ImplicitDifferentiation.PushforwardMul! — Type

PushforwardMul!{P,N}

Callable structure wrapping a pushforward with N-dimensional inputs into an in-place multiplication for vectors.

Fields

pushforward::P: the pushforward function

input_size::NTuple{N,Int}: the array size of the function input