API reference

Public

ImplicitDifferentiation.ImplicitFunction — Type

ImplicitFunction{F,C,L,B}

Differentiable wrapper for an implicit function defined by a forward mapping y and a set of conditions c.

The Jacobian of y(⋅) is computed using the implicit function theorem:

∂/∂y c(x, y(x)) * ∂y(x) = -∂/∂x c(x, y(x))

This requires solving a linear system A * J = -B.

Constructors

You can construct an ImplicitFunction from two callables (function-like objects) forward and conditions.

ImplicitFunction(
    forward, conditions;
    linear_solver=IterativeLinearSolver(), conditions_backend=nothing,
)

While forward does not not need to be compatible with automatic differentiation, conditions has to be (with the provided conditions_backend if there is one).

There are two possible signatures for forward and conditions, which must be consistent with one another:

1. Standard: `forward(x; kwargs...) = y` and `conditions(x, y; kwargs...) = c`
2. Byproduct: `forward(x; kwargs...) = (y, z)` and `conditions(x, y, z; kwargs...) = c`.

In both cases, x, y and c must be arrays with size(y) = size(c). In the second case, the byproduct z can be an arbitrary object generated by forward, but beware that we consider it constant for differentiation purposes.

Callable behavior

An ImplicitFunction object implicit behaves like a function, and every call is differentiable.

implicit(x::AbstractArray; kwargs...)

This returns exactly implicit.forward(x; kwargs...), which as we mentioned can be either an array y or a tuple (y, z).

Fields

forward::F

conditions::C

linear_solver::L<:AbstractLinearSolver

conditions_backend::B<:Union{Nothing,AbstractBackend}

Warning

At the moment, conditions_backend can only be nothing or AD.ForwardDiffBackend(). We are investigating why the other backends fail.

ImplicitDifferentiation.DirectLinearSolver — Type

DirectLinearSolver

An implementation of AbstractLinearSolver using the built-in backslash operator.

ImplicitDifferentiation.IterativeLinearSolver — Type

IterativeLinearSolver

An implementation of AbstractLinearSolver using Krylov.gmres.

Internals

ImplicitDifferentiation.AbstractLinearSolver — Type

AbstractLinearSolver

All linear solvers used within an ImplicitFunction must satisfy this interface.

It can be useful to roll out your own solver if you need more fine-grained control on convergence / speed / behavior in case of singularity. Check out the source code of IterativeLinearSolver and DirectLinearSolver for implementation examples.

Required methods

presolve(linear_solver, A, y): Returns a matrix-like object A for which it is cheaper to solve several linear systems with different vectors b (a typical example would be to perform LU factorization).

solve(linear_solver, A, b): Returns a vector x satisfying Ax = b. If the linear system has not been solved to satisfaction, every element of x should be a NaN of the appropriate floating point type.

ChainRulesCore.rrule — Function

rrule(rc, implicit, x; kwargs...)

Custom reverse rule for an ImplicitFunction, to ensure compatibility with reverse mode autodiff.

This is only available if ChainRulesCore.jl is loaded (extension), except on Julia < 1.9 where it is always available.

We compute the vector-Jacobian product Jᵀv by solving Aᵀu = v and setting Jᵀv = -Bᵀu. Keyword arguments are given to both implicit.forward and implicit.conditions.

Index

ImplicitDifferentiation.AbstractLinearSolver

ImplicitDifferentiation.DirectLinearSolver

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.IterativeLinearSolver

ChainRulesCore.rrule