API reference

Index

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.PullbackMul!

ImplicitDifferentiation.PushforwardMul!

Docstrings

ImplicitDifferentiation.ImplicitFunction — Type

ImplicitFunction{F,C,L}

Differentiable wrapper for an implicit function x -> y(x) whose output is defined by conditions F(x,y(x)) = 0.

More generally, we consider functions x -> (y(x),z(x)) and conditions F(x,y(x),z(x)) = 0, where z(x) is a byproduct considered constant for differentiation purposes. If x ∈ ℝⁿ and y ∈ ℝᵈ, then we need as many conditions as output dimensions: F(x,y,z) ∈ ℝᵈ. Thanks to these conditions, we can compute the Jacobian of y(⋅) using the implicit function theorem:

∂₂F(x,y(x),z(x)) * ∂y(x) = -∂₁F(x,y(x),z(x))

This amounts to solving a linear system A * J = -B, where A ∈ ℝᵈˣᵈ, B ∈ ℝᵈˣⁿ and J ∈ ℝᵈˣⁿ.

Fields:

forward::F: callable of the form x -> (ŷ(x),z(x)).

conditions::C: callable of the form (x,y,z) -> F(x,y,z)

linear_solver::L: callable of the form (A,b) -> u such that Au = b, must be taken from Krylov.jl

ImplicitDifferentiation.ImplicitFunction — Method

implicit(x[; kwargs...])
implicit(x, Val(true), [; kwargs...])

Make ImplicitFunction callable by applying implicit.forward.

The first (default) call signature only returns y(x), while the second returns (y(x), z(x)).

ImplicitDifferentiation.ImplicitFunction — Method

ImplicitFunction(forward, conditions, linear_solver, ::Val{byproduct} = Val(false))

Construct an ImplicitFunction with linear_solver as the linear solver used in the implicit differentiation. By default, ImplicitFunction(forward, conditions, linear_solver) will assume that the forward function returns a single output and that the conditions function accepts two arguments.

In some cases, it may be that a byproduct of the forward function can be used in the conditions function to save computation. To tell ImplicitDifferentiation that the forward function returns a second constant byproduct output and that the conditions function accepts that byproduct as a third argument, you can construct the implicit function using ImplicitFunction(forward, conditions, linear_solver, Val(true)).

ImplicitDifferentiation.ImplicitFunction — Method

ImplicitFunction(forward, conditions, ::Val{byproduct} = Val(false))

Construct an ImplicitFunction with Krylov.gmres as the default linear solver. By default, ImplicitFunction(forward, conditions) will assume that the forward function returns a single output and that the conditions function accepts two arguments.

In some cases, it may be that a byproduct of the forward function can be used in the conditions function to save computation. To tell ImplicitDifferentiation that the forward function returns a second constant byproduct output and that the conditions function accepts that byproduct as a third argument, you can construct the implicit function using ImplicitFunction(forward, conditions, Val(true)).

ImplicitDifferentiation.PullbackMul! — Type

PullbackMul!{P,N}

Callable structure wrapping a pullback with N-dimensional outputs into an in-place multiplication for vectors.

Fields

pullback::P: the pullback of the function

output_size::NTuple{N,Int}: the array size of the function output

ImplicitDifferentiation.PushforwardMul! — Type

PushforwardMul!{P,N}

Callable structure wrapping a pushforward with N-dimensional inputs into an in-place multiplication for vectors.

Fields

pushforward::P: the pushforward function

input_size::NTuple{N,Int}: the array size of the function input