API reference

Public

ImplicitDifferentiation.ImplicitFunction — Type

ImplicitFunction{FF,CC,LS}

Differentiable wrapper for an implicit function defined by a forward mapping and a set of conditions.

Constructors

You can construct an ImplicitFunction from a forward mapping f and conditions c, both of which must be callables (function-like objects). While f does not not need to be compatible with automatic differentiation, c has to be.

ImplicitFunction(
    f, c[, HandleByproduct()];
    linear_solver=IterativeLinearSolver()
)

Callable behavior

An ImplicitFunction object implicit behaves like a function, and every call to it is differentiable.

implicit(x::AbstractArray[, ReturnByproduct()]; kwargs...)

Details

By default, we assume that the forward mapping is x -> y(x) and the conditions are c(x,y(x)) = 0.

If HandleByproduct() is passed as an argument to the constructor, we assume instead that the forward mapping is x -> (y(x),z(x)) and the conditions are c(x,y(x),z(x)) = 0. In this case, z(x) can contain additional information generated by the forward mapping, but beware that we consider it constant for differentiation purposes.

Given x ∈ ℝⁿ and y ∈ ℝᵈ, we need as many conditions as output dimensions: c(x,y,z) ∈ ℝᵈ. We can then compute the Jacobian of y(⋅) using the implicit function theorem:

∂₂c(x,y(x),z(x)) * ∂y(x) = -∂₁c(x,y(x),z(x))

This requires solving a linear system A * J = -B, where A ∈ ℝᵈˣᵈ, B ∈ ℝᵈˣⁿ and J ∈ ℝᵈˣⁿ.

Fields

forward::FF: a wrapper of type Forward around the callable f

conditions::CC: a wrapper of type Conditions around the callable c

linear_solver::LS: an object subtyping AbstractLinearSolver.

ImplicitDifferentiation.DirectLinearSolver — Type

DirectLinearSolver

An implementation of AbstractLinearSolver using the built-in backslash operator.

ImplicitDifferentiation.IterativeLinearSolver — Type

IterativeLinearSolver

An implementation of AbstractLinearSolver using Krylov.gmres.

ImplicitDifferentiation.HandleByproduct — Type

HandleByproduct

Trivial struct specifying that the forward mapping and conditions handle a byproduct.

Used in the constructor for ImplicitFunction.

ImplicitDifferentiation.ReturnByproduct — Type

ReturnByproduct

Trivial struct specifying that we want to obtain a byproduct in addition to the solution.

Used when calling an ImplicitFunction.

ChainRulesCore.rrule — Function

rrule(rc, implicit, x[, ReturnByproduct()]; kwargs...)

Custom reverse rule for an ImplicitFunction, to ensure compatibility with reverse mode autodiff.

This is only available if ChainRulesCore.jl is loaded (extension), except on Julia < 1.9 where it is always available.

By default, this returns a single output y(x) with a pullback accepting a single cotangent dy.

If ReturnByproduct() is passed as an argument, this returns a couple of outputs (y(x),z(x)) with a pullback accepting a couple of cotangents (dy, dz) (remember that z(x) is not differentiated so its cotangent is ignored).

We compute the vector-Jacobian product Jᵀv by solving Aᵀu = v and setting Jᵀv = -Bᵀu (see ImplicitFunction for the definition of A and B). Keyword arguments are given to both implicit.forward and implicit.conditions.

Internals

ImplicitDifferentiation.Forward — Type

Forward{byproduct,F}

Callable wrapper for a forward mapping f::F, which ensures that a byproduct z(x) is always returned in addition to y(x).

The type parameter byproduct is a boolean stating whether or not f natively returns z(x).

ImplicitDifferentiation.Conditions — Type

Conditions{byproduct,C}

Callable wrapper for conditions c::C, which ensures that a byproduct z is always accepted in addition to x and y.

The type parameter byproduct is a boolean stating whether or not c natively accepts z.

ImplicitDifferentiation.AbstractLinearSolver — Type

AbstractLinearSolver

All linear solvers used within an ImplicitFunction must satisfy this interface.

It can be useful to roll out your own solver if you need more fine-grained control on convergence / speed / behavior in case of singularity. Check out the source code of IterativeLinearSolver and DirectLinearSolver for implementation examples.

Required methods

presolve(linear_solver, A, y): Returns a matrix-like object A for which it is cheaper to solve several linear systems with different vectors b (a typical example would be to perform LU factorization).

solve(linear_solver, A, b): Returns a vector x satisfying Ax = b. If the linear system has not been solved to satisfaction, every element of x should be a NaN of the appropriate floating point type.

ImplicitDifferentiation.PushforwardMul! — Type

PushforwardMul!{P,N}

Callable structure wrapping a pushforward with N-dimensional inputs into an in-place multiplication for vectors.

Fields

pushforward::P: the pushforward function

input_size::NTuple{N,Int}: the array size of the function input

ImplicitDifferentiation.PullbackMul! — Type

PullbackMul!{P,N}

Callable structure wrapping a pullback with N-dimensional outputs into an in-place multiplication for vectors.

Fields

pullback::P: the pullback of the function

output_size::NTuple{N,Int}: the array size of the function output

Index

ImplicitDifferentiation.AbstractLinearSolver

ImplicitDifferentiation.Conditions

ImplicitDifferentiation.DirectLinearSolver

ImplicitDifferentiation.Forward

ImplicitDifferentiation.HandleByproduct

ImplicitDifferentiation.ImplicitFunction

ImplicitDifferentiation.IterativeLinearSolver

ImplicitDifferentiation.PullbackMul!

ImplicitDifferentiation.PushforwardMul!

ImplicitDifferentiation.ReturnByproduct

ChainRulesCore.rrule